refactor: move require_cols decorator to utils to use more generally
